+++ src/hwrenderer/vd_scene.h
@@ -21,6 +21,14 @@ namespace vdoom
 		bool OtherPlane;
 	};
 
+	enum VdHwRenderFlags
+	{
+		VDHW_RENDERFLOOR = 1,
+		VDHW_RENDERCEILING = 2,
+		VDHW_RENDERALL = 7,
+		VDHW_PROCESSED = 8
+	};
+
 	struct VdHwSceneStats
 	{
 		unsigned int Flats;
@@ -40,6 +48,7 @@ namespace vdoom
 		};
 
 		VdHwScene();
+		~VdHwScene();
 
 		void Clear();
 		void CollectWorld();
@@ -47,10 +56,16 @@ namespace vdoom
 		const VdHwFlatCommand *GetFlats() const;
 		unsigned int GetFlatCount() const;
 		const VdHwSceneStats &GetStats() const;
+		unsigned char GetSubsectorRenderFlags(const subsector_t *subsector) const;
 
 	private:
 		bool AddFlat(const subsector_t *subsector, sector_t *planeSector, sector_t *textureSector, int plane, bool otherPlane);
 		bool AddOtherPlaneFlat(const subsector_t *subsector, sector_t *planeSector, int plane);
+		bool AddSectorFlat(sector_t *planeSector, sector_t *textureSector, int plane);
+		bool EnsureFlagBuffers();
+		void SetSubsectorRenderFlags(const subsector_t *subsector, unsigned char flags);
+		unsigned char GetSectorRenderFlags(sector_t *sector) const;
+		void SetSectorRenderFlags(sector_t *sector, unsigned char flags);
 		void AddSubsectorFlats(const subsector_t *subsector);
 		void CollectBspFlats(void *node, unsigned int depth);
 		void CollectFallbackFlats();
@@ -60,6 +75,10 @@ namespace vdoom
 		VdHwFlatCommand Flats[MaxFlats];
 		VdHwSceneStats Stats;
 		unsigned int FlatCount;
+		unsigned char *SubsectorRenderFlags;
+		unsigned int SubsectorRenderFlagCount;
+		unsigned char *SectorRenderFlags;
+		unsigned int SectorRenderFlagCount;
 	};
 }